create SPI lowlayer conf struct for service layer
