+++ .vscode/settings.json
@@ -6,6 +6,7 @@
     "gpio.h": "c",
     "system_timer.h": "c",
     "uart.h": "c",
-    "spi_irq_pc.h": "c"
+    "spi_irq_pc.h": "c",
+    "spi_serv.h": "c"
   }
 }
