+++ Users/Init_Drive/SPI/SPI_def.h
@@ -18,4 +18,32 @@ typedef enum{
   DIV_256 = 7,
 }Baudrate_mapping;
 
+typedef struct{
+    uint8_t* transmitt_buffer;
+    uint8_t* receive_buffer;
+    uint8_t transmitt_buffer_length;
+    uint8_t receive_buffer_length;
+    uint8_t transmitt_buffer_index;
+    uint8_t receive_buffer_index;
+}SPI_Data_buffer;
+
+
+typedef struct{
+  SPI_TypeDef* const SPI;
+  SPI_Data_buffer* const SPI_data_containter_str;
+}SPI_LL_conf_st;
+
+//Specific for STM32F103
+typedef enum{
+  SPI1 = 0,
+  SPI2 = 1,
+  MAX_SPI,
+}SPI_HANDLES
+
+/// Hardware Configurations
+extern SPI_LL_conf_st SPI_hardware_conf[MAX_SPI];
+
+extern SPI_Data_buffer SPI1_data_container_str; 
+extern SPI_Data_buffer SPI2_data_container_str; 
+
 #endif
