+++ Users/Service/SPI_serv.c
@@ -13,7 +13,7 @@
 SPI_Write()
 @Brief: function handles transmit info from Master to Slave using Interrupt mechanism
 */
-void SPI_Write(SPI_TypeDef* SPI, uint8_t* data_u8_array, uint8_t data_length_u8){
+void SPI_Write(const SPI_hardware_conf SPI_index, uint8_t* data_u8_array, uint8_t data_length_u8){
   /*
     Master transmit sequence:
     - The transmit sequence begins when a byte is written in the Tx Buffer.
@@ -33,14 +33,16 @@ void SPI_Write(SPI_TypeDef* SPI, uint8_t* data_u8_array, uint8_t data_length_u8)
     4. Wait until RXNE=1 and read the last received data.
     5. Wait until TXE=1 and then wait until BSY=0 before disabling the SPI.
   */
-  
-  SPI1_data_container_str.transmitt_buffer = data_u8_array;
-  SPI1_data_container_str.transmitt_buffer_length = data_length_u8;
-  SPI1_data_container_str.transmitt_buffer_index = 0;
+  SPI_typdef* const SPI_conf = SPI_hardware_conf[SPI_index].SPI; 
+  SPI_Data_buffer* const data_container_str = SPI_hardware_conf[SPI_index].SPI_data_containter_str; 
+
+  data_container_str.transmitt_buffer = data_u8_array;
+  data_container_str.transmitt_buffer_length = data_length_u8;
+  data_container_str.transmitt_buffer_index = 0;
 
   /*
     Enable SPI's SPI transmission and Transmit interrupt  
   */
-  SPI->CR1 |= SPI_CR1_SPE;
-  SPI->CR2 |= SPI_CR2_TXEIE;
+  SPI_conf->CR1 |= SPI_CR1_SPE;
+  SPI_conf->CR2 |= SPI_CR2_TXEIE;
 }
